# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API VèoFood",
        "description": "Tài liệu API cho Đồ án VèoFood - Nhóm 15.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "VèoFood API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Món Ăn (Products)"
                ],
                "summary": "Lấy danh sách tất cả món ăn",
                "description": "Trả về mảng dữ liệu món ăn kèm category",
                "operationId": "getProductsList",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm theo tên món ăn",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "phở"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "Món Ăn (Products)"
                ],
                "summary": "Tạo món ăn mới",
                "description": "Thêm món ăn",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Trà sữa"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 30000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Món Ăn (Products)"
                ],
                "summary": "Lấy chi tiết món ăn theo id",
                "description": "Trả về thông tin một món ăn kèm category",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy món ăn"
                    }
                }
            },
            "put": {
                "tags": [
                    "Món Ăn (Products)"
                ],
                "summary": "Cập nhật món ăn",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Món Ăn (Products)"
                ],
                "summary": "Xóa món ăn",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Nhập Bearer token để xác thực",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    }
}